backend/auth-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pyjwt
bcrypt